// frontend/src/lib/api/generated/models/projectResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { PlatformIdentityPayload } from "./platformIdentityPayload.ts";

export interface ProjectResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  created_at: string;
  default_branch?: string;
  display_name: string;
  id: string;
  local_path: string;
  platform_identity?: PlatformIdentityPayload;
  updated_at: string;
}