// frontend/src/lib/api/generated/models/worktreeResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface WorktreeResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  branch: string;
  created_at: string;
  id: string;
  is_hidden: boolean;
  is_primary: boolean;
  linked_issue_numbers: number[];
  path: string;
  project_id: string;
  session_backend: string;
  updated_at: string;
}